Przeniesienie klasy Node do wnetrza klasy StackList - usuniecie zbednej klasy, korekta modyfikatora final
