+++ src/main/java/edu/kis/vh/stacks/list/StackList.java
@@ -38,5 +38,38 @@ public class StackList {
         last = last.getPrevious();
         return ret;
     }
+    
+    class Node {
 
+    	// utworzone zostalo pole klasy Node oraz przy wkladaniu elementu do list
+    	// tworzona jest jej instancja jesli pole nie jest zainicjalizowane
+    	// dostep do poprzedniego i nastepnego pola realizowany jest przez gettery
+    	private final int value;
+    	private Node previous;
+    	private Node next;
+
+    	public Node(int i) {
+    		value = i;
+    	}
+
+    	int getValue() {
+    		return value;
+    	}
+
+    	Node getNext() {
+    		return next;
+    	}
+
+    	void setNext(Node next) {
+    		this.next = next;
+    	}
+
+    	Node getPrevious() {
+    		return previous;
+    	}
+
+    	void setPrevious(Node previous) {
+    		this.previous = previous;
+    	}
+    }
 }
